+++ CHANGELOG.md
@@ -44,12 +44,16 @@
 * `isModifierDescriptor` is no longer exposed
 * `deepEqual` is no longer exposed, use `comparer.structural` instead
 
-### extenddObservable
+### extendObservable
 
 * killed using decorators as functions on values
 * `extendObservable` can no longer be used to redefine or update existing properties! To declare existing properties to be observable use 'decorate'
 * `extendObservable(x, { y: computed(...)})` => `extendObservable(x, { get y() {}})` or `extendObservable(x, { get y() {}}, { y: computed(opts)})`
 * `action.bound` => use decorator
+
+#### observable
+
+* `observable.shallowX` is deprecated, use `observable.X(value, { deep: false })` instead
 
 ## Non breaking changes